text_to_sql/prototype/requirements.txt
Flask
python-dotenv
langchain
langchain-openai
langchain-chroma
langchain-huggingface
langchain-cohere
chromadb
langgraph
pydantic
pandas